version 0.2
After the game, the king and the pawn go into the same box. Nothing happened to anyone.

+++ Endgame/Assets/Scripts/Fade.cs
@@ -36,7 +36,7 @@ public class Fade : MonoBehaviour
                 isFading = false;
                 return;
             }
-            float num = frames / maxFrames;
+            float num = (float)(frames / maxFrames);
             if (!fadingOut)
             {
                 num = 1f - num;
